test: move state-input test to MIL passes test file
Moved test_handle_unused_inputs_skips_state from the NN backend test
file to the MIL passes test file, as requested in code review.

The fix itself is in the NN backend pass, but the test belongs with
the MIL-level pass tests since it validates MIL program behavior.
